+++ utils/checks.py
@@ -56,6 +56,23 @@ def is_channel_mod():
     return commands.check(predicate)
 
 
+def is_channel_mod_somewhere():
+    """Checks if the author has manage guild permissions in any guild"""
+    async def predicate(ctx):
+        ret = await is_owner_check(ctx)
+        if ret:
+            return True
+        if ctx.guild is not None:
+            return await check_guild_permissions(ctx, {'manage_channels': True})
+        for guild in ctx.bot.get_user_guilds(ctx.author.id):
+            member = guild.get_member(ctx.author.id)
+            permissions = member.guild_permissions
+            if permissions.administrator or permissions.manage_channels:
+                return True
+        return False
+    return commands.check(predicate)
+
+
 def is_tracking_world():
     """Checks if the current server is tracking a tibia world