+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## Version 1.6.1 (2018-09-06)
+- Fixed bug with /worlds
+- Unified missing argument error messages
+- Commands in `/help` are now sorted alphabetically
+
 ## Version 1.6.0 (2018-08-27)
 - New command: `/rashid`.
 - `/time` no longer displays Mexico and Brazil's timezones.
@@ -210,7 +215,7 @@
 - Migrated many services from Tibia.com to TibiaData.com for better reliability.
 - TibiaWiki database is now more recent and is now a [separate project](https://github.com/Galarzaa90/tibiawiki-sql)
 - Added tons of new commands and rewrote many of them.
-- Added [documentation site](https://galarzaa90.github.io/NabBot/)
+- Added [documentation site](https://nabdev.github.io/NabBot/)
 - And too many changes too list them here.
 
 ## Version 0.1.3 (2018-03-08)